+++ CHANGELOG.md
@@ -17,6 +17,8 @@
   * MPEditElementAux.separateLineSegmentsPerType() moved to TH2FileEditElementEditController.groupLineSegmentsForSimplification().
   * THHasOptionsMixin constructors now receive optionsMap and attrOptionsMap parameters to initialize the element with existing options maps.
   * Identifying methods and classes actually used in line simplification with a 'MPSimplification' prefix in their names.
+  * Moved fromExisting constructors from individual MPCommand classes to MPCommandFactory:
+    * areaBorderTHID.
 
 ## 0.2.23 - 2025-11-27 - The Peruaçu release
 * New features: